pyproject.toml
[project]
name = "hcai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.2",
    "langchain>=1.1.0",
    "langchain-ollama>=1.0.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "openai-whisper>=20250625",
    "pyaudio",
    "scikit-learn>=1.7.2",
]
